+++ .changeset/clever-glasses-tickle.md
@@ -1,0 +1,5 @@
+---
+"claude-dev": patch
+---
+
+Moved rule file conversions
