--- src/shared/proto-conversions/index.ts
@@ -1,1 +1,0 @@
-export * from "./file"
